% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cmd.HPTN071.simulator}
\alias{cmd.HPTN071.simulator}
\title{Command line generator for \code{HPTN071.simulator}}
\usage{
cmd.HPTN071.simulator(outdir, seed = NA, opt.acute = "high",
  opt.intervention = "fast")
}
\value{
command line string
}
\description{
Command line generator for \code{HPTN071.simulator}
}
\examples{
##--------------------------------------------------------------------------------------------------------
##	create command line string to simulate epidemic scenarios under the 
##	regional HPTN071/PopART model
##	code by Mike Pickles and Anne Cori, version 150119
##--------------------------------------------------------------------------------------------------------
cat(cmd.HPTN071.simulator('.', seed=NA, opt.acute='high', opt.intervention='fast'))
}

