% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PANGEA.TransmissionEdgeEvolutionaryRate.create.sampler}
\alias{PANGEA.TransmissionEdgeEvolutionaryRate.create.sampler}
\title{Create sampler of Evolutionary Rates for transmission edges}
\usage{
PANGEA.TransmissionEdgeEvolutionaryRate.create.sampler(er.meanlog, er.sdlog)
}
\arguments{
\item{er.shift}{shift to mean of the log normal density}
}
\value{
R function
}
\description{
Returns a function to draw evolutionary rates for transmission edges. Currently modelled with a log normal density.
}

