\docType{data}
\name{PR.PACKAGE}
\alias{PR.PACKAGE}
\title{This file contains R functions that provide an interface to HPC scripting
Shell scripts are generated that can be either run directly, or submitted to an HPC system.}
\format{chr "rPANGEAHIVsim"}
\usage{
  PR.PACKAGE
}
\description{
  This file contains R functions that provide an interface
  to HPC scripting Shell scripts are generated that can be
  either run directly, or submitted to an HPC system.
}
\keyword{datasets}

