% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rPANGEAHIVsim.pipeline.args}
\alias{rPANGEAHIVsim.pipeline.args}
\title{Arguments for rPANGEAHIV simulation pipeline}
\usage{
rPANGEAHIVsim.pipeline.args(yr.start = 1980, yr.end = 2020, seed = 42,
  s.INC.recent = 0.1, s.INC.recent.len = 5, s.PREV.min = 0.01,
  s.PREV.max = 0.25, s.PREV.base = exp(1), s.INTERVENTION.start = 2015,
  s.INTERVENTION.mul = 2, s.ARCHIVAL.n = 50, s.MODEL = "Prop2DiagB4I",
  s.PREV.max.n = NA, s.INTERVENTION.prop = NA, epi.model = "HPTN071",
  epi.acute = "high", epi.intervention = "fast", epi.dt = 1/48,
  epi.import = 0.1, root.edge.fixed = 0, v.N0tau = 35800, v.r = 2,
  v.T50 = -1, wher.mu = 0.005, wher.sigma = 0.8, bwerm.mu = 1.5,
  bwerm.sigma = 0.12, er.gamma = 4, sp.prop.of.sexactive = 0.05,
  report.prop.recent = 0.2, dbg.GTRparam = 0, dbg.rER = 0,
  index.starttime.mode = "normal", startseq.mode = "many",
  seqtime.mode = "Gamma9")
}
\arguments{
\item{yr.start}{Start year of the epi simulation (default 1980)}

\item{yr.end}{First year after the epi simulation (default 2020)}

\item{seed}{Random number seed}

\item{s.INC.recent}{Proportion of incident cases sampled (not used)}

\item{s.INC.recent.len}{Number of last year in which an exact proportion of incident cases is sampled (not used)}

\item{s.PREV.min}{Proportion of infected cases sampled at start of the simulation}

\item{s.PREV.max}{Proportion of infected cases sampled at the end of the simulation}

\item{s.INTERVENTION.start}{Year in which the community intervention starts}

\item{s.INTERVENTION.mul}{Multiplier of number of sequences sampled per year after start of the intervention}

\item{s.ARCHIVAL.n}{Total number of sequences sampled before diagnosis}

\item{s.MODEL}{Sampling model to use}

\item{s.PREV.max.n}{Number of infected cases sampled (usually NA, only used for Prop2Untreated)}

\item{s.INTERVENTION.prop}{Proportion of infected cases that are sampled from after intervention start}

\item{epi.model}{The epi model used to create the epi simulation (default HPTN071, alternatively DSPS)}

\item{epi.dt}{Time increment of the epi simulation (default 1/48)}

\item{epi.import}{Proportion of imported cases of all cases (default 0.1)}

\item{root.edge.fixed}{Boolean; fix evolutionary rate of root edges to mean rate if true}

\item{v.N0tau}{Parameter of BEAST::LogisticGrowthN0::N0 (default 3.58e4)}

\item{v.r}{Parameter of BEAST::LogisticGrowthN0::r (default 2)}

\item{v.T50}{Parameter of BEAST::LogisticGrowthN0::T50}

\item{wher.mu}{Mean within host evolutionary rate of log normal density}

\item{wher.sigma}{Standard deviation in within host evolutionary rate of log normal density}

\item{bwerm.mu}{Mean between host evolutionary rate multiplier of log normal density}

\item{bwerm.sigma}{Standard deviation in between host evolutionary rate multiplier of log normal density}

\item{sp.prop.of.sexactive}{Proportion of population sampled in seroprevalence survey}

\item{report.prop.recent}{Proportion of individuals for whom recency of infection should be reported}

\item{dbg.GTRparam}{debug flag}

\item{dbg.rER}{debug flag}

\item{index.starttime.mode}{distribution to sample times for starting sequence: Normal(1960,7) or Unif(1959.75, 1960.25)}

\item{startseq.mode}{Number of different starting sequences either 'many' or 'one'.}
}
\value{
data.table
}
\description{
Construct table with all input arguments to the rPANGEAHIV simulation pipeline.
The demographic within host coealescent model is N0tau*(1+exp(-r*T50))/(1+exp(-r*(T50-t))). Default parameters are set
so that the curve asymptotes at Ne*tau=3e5 and reaches half its asymptotic value 1 year post infection. Branch lengths are
multiplied by within host evolutionary rates, and within host branch lengths ending in a transmission event are multiplied
in addition with a between-host evolutionary rate multiplier.
}
\examples{
##--------------------------------------------------------------------------------------------------------
##	example pipeline to simulate sequences for a given HPTN071 epi simulation  
##--------------------------------------------------------------------------------------------------------
\dontrun{
#	re-name the following:
outdir			<- '/Users/Oliver/duke/2015_various'
#	input arguments for the pipeline
pipeline.args	<- rPANGEAHIVsim.pipeline.args( yr.start=1985, yr.end=2020, seed=42, s.MODEL='Fixed2Prop', report.prop.recent=1.0,
			s.PREV.max.n=1600, s.INTERVENTION.prop=0.5, s.INTERVENTION.start=2015, s.INTERVENTION.mul=NA, s.ARCHIVAL.n=50,
			epi.model='HPTN071', epi.acute='high', epi.intervention='fast', epi.dt=1/48, epi.import=0.05, root.edge.fixed=0,
			v.N0tau=1, v.r=2.851904, v.T50=-2,
			wher.mu=log(0.00447743)-0.5^2/2, wher.sigma=0.5, bwerm.mu=log(0.002239075)-0.3^2/2, bwerm.sigma=0.3, er.gamma=4,
			dbg.GTRparam=0, dbg.rER=0, index.starttime.mode='fix1970', startseq.mode='one', seqtime.mode='AtDiag')									
#	
#	call simulation pipeline
#	this generates a UNIX batch file if no HPC system is detected, or
#	this generates and runs a qsub file if an HPC system is detected 
#
file			<- rPANGEAHIVsim.pipeline(outdir, pipeline.args=pipeline.args)
cat(file)
}
}

