% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rPANGEAHIVsim.pipeline}
\alias{rPANGEAHIVsim.pipeline}
\title{rPANGEAHIV simulation pipeline}
\usage{
rPANGEAHIVsim.pipeline(outdir, pipeline.args = rPANGEAHIVsim.pipeline.args())
}
\arguments{
\item{pipeline.args}{Input arguments for the simulation}

\item{indir}{Input directory}

\item{infile.ind}{Input file with individual metavariables}

\item{infile.trm}{Input file with transmission events}
}
\value{
file name of qsub or UNIX batch file
}
\description{
Reads two files \code{infile.ind} and \code{infile.trm} from the epi simulator in directory \code{indir} and produces a UNIX batch
file that contains all the simulation steps in directory \code{outdir}.
}
\examples{
##--------------------------------------------------------------------------------------------------------
##	example pipeline to simulate sequences for a given HPTN071 epi simulation  
##--------------------------------------------------------------------------------------------------------
\dontrun{
#	re-name the following:
outdir			<- '/Users/Oliver/duke/2015_various'
#	input arguments for the pipeline
pipeline.args	<- rPANGEAHIVsim.pipeline.args( yr.start=1985, yr.end=2020, seed=42, s.MODEL='Fixed2Prop', report.prop.recent=1.0,
			s.PREV.max.n=1600, s.INTERVENTION.prop=0.5, s.INTERVENTION.start=2015, s.INTERVENTION.mul=NA, s.ARCHIVAL.n=50,
			epi.model='HPTN071', epi.acute='high', epi.intervention='fast', epi.dt=1/48, epi.import=0.05, root.edge.fixed=0,
			v.N0tau=1, v.r=2.851904, v.T50=-2,
			wher.mu=log(0.00447743)-0.5^2/2, wher.sigma=0.5, bwerm.mu=log(0.002239075)-0.3^2/2, bwerm.sigma=0.3, er.gamma=4,
			dbg.GTRparam=0, dbg.rER=0, index.starttime.mode='fix1970', startseq.mode='one', seqtime.mode='AtDiag')									
#	
#	call simulation pipeline
#	this generates a UNIX batch file if no HPC system is detected, or
#	this generates and runs a qsub file if an HPC system is detected 
#
file			<- rPANGEAHIVsim.pipeline(outdir, pipeline.args=pipeline.args)
cat(file)
}
##--------------------------------------------------------------------------------------------------------
##	example pipeline to simulate sequences for a given DSPS epi simulation  
##--------------------------------------------------------------------------------------------------------
\dontrun{
indir			<- system.file(package="rPANGEAHIVsim", "misc")
#	re-name the following:
tmpdir			<- '/Users/Oliver/git/HPTN071sim/tmp140912'
dir.create(tmpdir, showWarnings=FALSE)
#	simulation input files from the epi-simulator
infile.trm		<- '140911_DSPS_RUN002_TRM.csv'
infile.ind		<- NULL
#	input arguments for the pipeline
pipeline.args	<- rPANGEAHIVsim.pipeline.args( yr.start=1980, yr.end=2020, seed=42, s.PREV.min=0.01, s.PREV.max=0.70, epi.model='DSPS', epi.dt=1/48, epi.import=0.1 )	
infile.args		<- paste(tmpdir,'/',substr(infile.trm, 1, nchar(infile.trm)-7), 'PipeArgs.R',sep='')
save(pipeline.args, file=infile.args)
#	
#	call simulation pipeline
#	this generates a UNIX batch file if no HPC system is detected, or
#	this generates and runs a qsub file if an HPC system is detected 
#
file			<- rPANGEAHIVsim.pipeline(indir, infile.ind, infile.trm, tmpdir, pipeline.args=pipeline.args)
cat(file)
}
}

