% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hptn071.prog.R
\name{sim.regional.args}
\alias{sim.regional.args}
\title{Input arguments to simulate sequences and viral phylogenetic trees\cr under the Regional Transmission and Intervention Model}
\usage{
sim.regional.args(yr.start = 1985, yr.end = 2020, seed = 42,
  s.INC.recent = NA, s.INC.recent.len = NA, s.PREV.min = NA,
  s.PREV.max = NA, s.PREV.base = NA, s.INTERVENTION.start = 2015,
  s.INTERVENTION.mul = NA, s.ARCHIVAL.n = 50, s.MODEL = "Fixed2Prop",
  s.PREV.max.n = 1600, s.INTERVENTION.prop = 0.5, epi.model = "HPTN071",
  epi.acute = "high", epi.intervention = "fast", epi.dt = 1/48,
  epi.import = 0.05, root.edge.fixed = 0, v.N0tau = 1, v.r = 2.851904,
  v.T50 = -2, wher.mu = log(0.00447743) - 0.5^2/2, wher.sigma = 0.5,
  bwerm.mu = log(0.002239075) - 0.3^2/2, bwerm.sigma = 0.3, er.gamma = 4,
  er.gtr = "GTR_VARIABLE", sp.prop.of.sexactive = 0.05,
  report.prop.recent = 1, dbg.GTRparam = 0, dbg.rER = 0,
  index.starttime.mode = "fix1970", startseq.mode = "one",
  seqtime.mode = "AtDiag")
}
\arguments{
\item{yr.start}{Start year of the epi simulation (default: 1985)}

\item{yr.end}{First year after the epi simulation (default: 2020; optional 2018)}

\item{seed}{Random number seed (default: 42; optional integer)}

\item{s.INC.recent}{Deprecated}

\item{s.INC.recent.len}{Deprecated}

\item{s.PREV.min}{Deprecated}

\item{s.PREV.max}{Deprecated}

\item{s.INTERVENTION.start}{Year in which the community intervention starts (default: 2015). See \url{https://github.com/olli0601/PANGEA.HIV.sim#sequence-sampling-model}.}

\item{s.INTERVENTION.mul}{Deprecated}

\item{s.ARCHIVAL.n}{Total number of sequences sampled at random from infected individuals before 2000 (default: 50). See \url{https://github.com/olli0601/PANGEA.HIV.sim#sequence-sampling-model}.}

\item{s.MODEL}{Sampling model to use (default: 'Fixed2Prop'). See \url{https://github.com/olli0601/PANGEA.HIV.sim#sequence-sampling-model}.}

\item{s.PREV.max.n}{Number of infected cases sampled (default: 1600; optional 3200). See \url{https://github.com/olli0601/PANGEA.HIV.sim#sequence-sampling-model}.}

\item{s.INTERVENTION.prop}{Proportion of infected cases that are sampled from after intervention start (default: 0.5; optional 0.85). See \url{https://github.com/olli0601/PANGEA.HIV.sim#sequence-sampling-model}.}

\item{epi.model}{The epi model used to create the epi simulation (default: 'HPTN071')}

\item{epi.acute}{The proportion of transmissions in 2015 from individuals in their first three months of infection (default: 'high'/40\%, optional: 'low'/10\%)}

\item{epi.intervention}{The scale of the intervention package (default:'fast', optional: 'slow', 'none')}

\item{epi.dt}{Time increment of the epi simulation (default: 1/48, this is fixed to the underlying epidemic simulation)}

\item{epi.import}{Proportion of viral introductions amongst new cases per year (default: 0.05, optional: 0.2). See \url{https://github.com/olli0601/PANGEA.HIV.sim#viral-introductions}.}

\item{root.edge.fixed}{Fix evolutionary rate of root edges of new lineages in the population to mean evolutionary rate (default: 1, optional: 0). See \url{https://github.com/olli0601/PANGEA.HIV.sim#overall-evolutionary-rates-of-transmission-and-non-transmission-lineages}.}

\item{v.N0tau}{Within host effective population size at time of infection (default 1). See \url{https://github.com/olli0601/PANGEA.HIV.sim#dated-viral-phylogenies}.}

\item{v.r}{Parameter logistic growth model (default 2.851904). See \url{https://github.com/olli0601/PANGEA.HIV.sim#dated-viral-phylogenies}.}

\item{v.T50}{Parameter logistic growth model (default -2). See \url{https://github.com/olli0601/PANGEA.HIV.sim#dated-viral-phylogenies}.}

\item{wher.mu}{Mean of within host evolutionary rate (default: log(0.00447743)-0.5^2/2). See \url{https://github.com/olli0601/PANGEA.HIV.sim#overall-evolutionary-rates-of-transmission-and-non-transmission-lineages}.}

\item{wher.sigma}{Standard deviation of within host evolutionary (default: 0.5). See \url{https://github.com/olli0601/PANGEA.HIV.sim#overall-evolutionary-rates-of-transmission-and-non-transmission-lineages}.}

\item{bwerm.mu}{Mean of between host evolutionary rate (default: log(0.002239075)-0.3^2/2). See \url{https://github.com/olli0601/PANGEA.HIV.sim#overall-evolutionary-rates-of-transmission-and-non-transmission-lineages}.}

\item{bwerm.sigma}{Standard deviation of between host evolutionary rate (default: 0.3). See \url{https://github.com/olli0601/PANGEA.HIV.sim#overall-evolutionary-rates-of-transmission-and-non-transmission-lineages}.}

\item{er.gamma}{Site heterogeneity parameter (default: 4, optional 0)}

\item{er.gtr}{Evolution model identifier. Default: GTR_VARIABLE. Alternatively: GTR_POL_CP1.}

\item{sp.prop.of.sexactive}{Proportion of population sampled in seroprevalence survey (default: 0.05)}

\item{report.prop.recent}{Proportion of individuals in seroprevalence survey for whom recency of infection should be reported (default: 1)}

\item{dbg.GTRparam}{Use mean GTR rates instead of samples from empirical prior (default: 0, optional 1)}

\item{dbg.rER}{Use mean evolutionary rate instead of samples from empirical prior (default: 0, optional 1)}

\item{index.starttime.mode}{Sampling distribution of sampling times of starting sequences (default: 'fix1970', optional 'fix19XX' where XX between 45-79, 'normal' meaning Normal(1955,7)). See \url{https://github.com/olli0601/PANGEA.HIV.sim#seed-sequences}.}

\item{startseq.mode}{Number of different starting sequences (default: 'one', optional 'many'). See \url{https://github.com/olli0601/PANGEA.HIV.sim#seed-sequences}.}

\item{seqtime.mode}{Time at which sequences are sampled (default: 'AtDiag'). See \url{https://github.com/olli0601/PANGEA.HIV.sim#sequence-sampling-model}.}
}
\value{
data.table with columns 'stat' (name of input argument) and 'v' (value of input argument).
}
\description{
Set all input arguments to the simulation. Please see for model details \url{https://github.com/olli0601/PANGEA.HIV.sim}
}
\seealso{
\code{\link{sim.regional}}
}

