% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PANGEA.WithinHostEvolutionaryRate.create.sampler.v1}
\alias{PANGEA.WithinHostEvolutionaryRate.create.sampler.v1}
\title{Create sampler of Within Host Evolutionary Rates}
\usage{
PANGEA.WithinHostEvolutionaryRate.create.sampler.v1(wher.mu = log(0.005),
  wher.sigma = 0.8)
}
\arguments{
\item{wher.mu}{mean of the log normal density}

\item{wher.sigma}{standard deviation of the log normal density}
}
\value{
R function
}
\description{
Returns a function to draw within host evolutionary rates. Currently modelled with a log normal density.
}

